+++ app/forms/mappings/Formatters.scala
@@ -223,8 +223,8 @@ trait Formatters {
     invalidCharactersKey: String = "error.invalidCharacters",
     invalidFormatKey: String = "error.invalidFormat",
     invalidValueKey: String = "error.invalidValue"
-  ): Formatter[Double] =
-    new Formatter[Double] {
+  ): Formatter[BigDecimal] =
+    new Formatter[BigDecimal] {
 
       private val invalidCharactersRegex = """^[0-9.]*$"""
       private val invalidFormatRegex     = """^[0-9]*(\.[0-9]{1,2})?$"""
@@ -232,7 +232,7 @@ trait Formatters {
 
       private val baseFormatter = stringFormatter(requiredKey)
 
-      override def bind(key: String, data: Map[String, String]): Either[Seq[FormError], Double] =
+      override def bind(key: String, data: Map[String, String]): Either[Seq[FormError], BigDecimal] =
         baseFormatter
           .bind(key, data)
           .right
@@ -243,10 +243,10 @@ trait Formatters {
             case s if !s.matches(invalidCharactersRegex) => Left(Seq(FormError(key, invalidCharactersKey)))
             case s if !s.matches(invalidFormatRegex)     => Left(Seq(FormError(key, invalidFormatKey)))
             case s if !s.matches(invalidValueRegex)      => Left(Seq(FormError(key, invalidValueKey)))
-            case s                                       => Right(s.toDouble)
+            case s                                       => Right(BigDecimal(s))
           }
 
-      override def unbind(key: String, value: Double): Map[String, String] =
-        baseFormatter.unbind(key, value.toString)
+      override def unbind(key: String, value: BigDecimal): Map[String, String] =
+        baseFormatter.unbind(key, value.toString())
     }
 }
